archive/legacy_attempt/requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.22.0
Pillow>=10.0.0
scipy>=1.11.0
matplotlib>=3.8.0
Flask>=3.0.0
Flask-Cors>=4.0.0
scikit-learn>=1.3.0